requirements.txt
requests>=2.31.0
numpy>=1.26.0
opencv-python>=4.9.0.80
Pillow>=10.2.0
Flask>=3.0.0
pytesseract>=0.3.10
